pyproject.toml
[build-system]
requires = ["setuptools>=65.5.3"]
build-backend = "setuptools.build_meta"

[project]
name = "stableviews"
version = "0.0.0"
requires-python = ">=3.11"
dependencies = [
    "roma",
    "viser",
    "tyro",
    "ninja",
    "gradio",
    "einops",
    "colorama",
    "splines",
    "kornia",
    "open-clip-torch",
    "diffusers",
    "numpy",
    "imageio[ffmpeg]",
    "huggingface-hub"
]

[project.optional-dependencies]
dev = ["ruff", "ipdb", "pytest", "line_profiler", "pre-commit"]

[tool.setuptools.packages.find]
include = ["stableviews"]

[tool.pyright]
extraPaths = ["third_party/dust3r"]
